+++ offchain/eth-input-reader/src/setup.rs
@@ -1,88 +1,107 @@
 // (c) Cartesi and individual authors (see AUTHORS)
 // SPDX-License-Identifier: Apache-2.0 (see LICENSE)
 
-use eth_state_client_lib::{
-    config::SCConfig, error::StateServerError, BlockServer,
-    GrpcStateFoldClient, StateServer,
-};
+use eth_block_history::{BlockArchive, BlockSubscriber, SubscriptionError};
+use eth_state_fold::StateFoldEnvironment;
 use eth_state_fold_types::BlockStreamItem;
+use ethers::providers::{
+    Http, HttpRateLimitRetryPolicy, Provider, RetryClient,
+};
 use rollups_events::DAppMetadata;
 use snafu::ResultExt;
-use tokio_stream::{Stream, StreamExt};
-use tonic::transport::Channel;
-use types::foldables::authority::{RollupsInitialState, RollupsState};
+use std::sync::{Arc, Mutex};
+use tokio_stream::Stream;
+use types::UserData;
+use url::Url;
 
+use crate::error::ParseSnafu;
 use crate::{
     config::EthInputReaderConfig,
-    error::{
-        BrokerSnafu, ChannelSnafu, ConnectSnafu, EthInputReaderError,
-        StateServerSnafu,
-    },
+    error::{BrokerSnafu, EthInputReaderError},
     machine::{BrokerStatus, Context},
     metrics::EthInputReaderMetrics,
 };
 
-const BUFFER_LEN: usize = 256;
+pub type InputProvider = Provider<RetryClient<Http>>;
 
-pub async fn create_state_server(
-    config: &SCConfig,
-) -> Result<
-    impl StateServer<InitialState = RollupsInitialState, State = RollupsState>
-        + BlockServer,
-    EthInputReaderError,
-> {
-    let channel = Channel::from_shared(config.grpc_endpoint.to_owned())
-        .context(ChannelSnafu)?
-        .connect()
-        .await
-        .context(ConnectSnafu)?;
+pub async fn create_provider(
+    config: &EthInputReaderConfig,
+) -> Result<Arc<InputProvider>, EthInputReaderError> {
+    let http = Http::new(
+        Url::parse(&config.bh_config.http_endpoint).context(ParseSnafu)?,
+    );
 
-    Ok(GrpcStateFoldClient::new_from_channel(channel))
+    let retry_client =
+        RetryClient::new(http, Box::new(HttpRateLimitRetryPolicy), 10, 1000);
+
+    let provider = Provider::new(retry_client);
+
+    Ok(Arc::new(provider))
 }
 
-pub async fn create_block_subscription(
-    client: &impl BlockServer,
-    confirmations: usize,
+pub async fn create_block_subscriber(
+    config: &EthInputReaderConfig,
+    provider: Arc<InputProvider>,
+) -> Result<Arc<BlockSubscriber<InputProvider>>, EthInputReaderError> {
+    Ok(Arc::new(
+        BlockSubscriber::start(
+            Arc::clone(&provider),
+            config.bh_config.ws_endpoint.to_owned(),
+            config.bh_config.block_timeout,
+            config.bh_config.max_depth,
+        )
+        .await
+        .expect("should create block history"),
+    ))
+}
+
+pub async fn create_subscription(
+    block_subscriber: Arc<BlockSubscriber<InputProvider>>,
 ) -> Result<
-    impl Stream<Item = Result<BlockStreamItem, StateServerError>>
-        + Send
+    impl Stream<Item = Result<BlockStreamItem, SubscriptionError<InputProvider>>>
         + std::marker::Unpin,
     EthInputReaderError,
 > {
-    let s = client
-        .subscribe_blocks(confirmations)
+    Ok(block_subscriber
+        .subscribe_new_blocks_at_depth(10)
         .await
-        .context(StateServerSnafu)?;
+        .expect("should subscribe new blocks"))
+}
 
-    let s = {
-        use futures::StreamExt;
-        s.ready_chunks(BUFFER_LEN)
-    };
-
-    let s = s.filter_map(
-        |mut x| {
-            if x.len() == BUFFER_LEN {
-                None
-            } else {
-                x.pop()
-            }
-        },
+pub async fn create_env(
+    config: &EthInputReaderConfig,
+    provider: Arc<InputProvider>,
+    block_archive: Arc<BlockArchive<InputProvider>>,
+) -> Result<
+    Arc<StateFoldEnvironment<InputProvider, Mutex<UserData>>>,
+    EthInputReaderError,
+> {
+    let env = StateFoldEnvironment::new(
+        provider,
+        Some(block_archive),
+        config.sf_config.safety_margin,
+        config.sf_config.genesis_block,
+        config.sf_config.query_limit_error_codes.clone(),
+        config.sf_config.concurrent_events_fetch,
+        10000,
+        Mutex::new(UserData::default()),
     );
 
-    Ok(s)
+    Ok(Arc::new(env))
 }
 
 pub async fn create_context(
     config: &EthInputReaderConfig,
-    block_server: &impl BlockServer,
+    block_subscriber: Arc<BlockSubscriber<InputProvider>>,
     broker: &impl BrokerStatus,
     dapp_metadata: DAppMetadata,
     metrics: EthInputReaderMetrics,
 ) -> Result<Context, EthInputReaderError> {
-    let genesis_timestamp: u64 = block_server
-        .query_block(config.dapp_deployment.deploy_block_hash)
+    let genesis_timestamp: u64 = block_subscriber
+        .block_archive
+        .block_with_hash(&config.dapp_deployment.deploy_block_hash)
         .await
-        .context(StateServerSnafu)?
+        .expect("should get genesis block")
         .timestamp
         .as_u64();
     let epoch_length = config.epoch_duration;